(chore) regex: extract pattern analysis and replacement logic from Matcher
Split Matcher.java (2354 lines) into focused helper classes to improve
maintainability while preserving the public API surface unchanged.

Extract MatcherPatternAnalysis (257 lines): static pattern analysis
methods for anchoring bounds, hit-end, and require-end heuristics.

Extract MatcherReplacementProcessor (168 lines): replacement string
processing with group reference resolution.

Matcher.java reduced to 2025 lines (-329 lines, -14%).

Closes #335
